requirements.txt
aiogram==2.20
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.5
Deprecated==1.2.13
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.1
tomli==2.0.1
typing_extensions==4.2.0
wrapt==1.14.1
yarl==1.7.2
